apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dental-clinic-api"
version = "0.1.0"
description = "Production-oriented FastAPI starter for the dental clinic platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg2-binary",
    "pydantic-settings",
    "python-dotenv",
    "google-genai>=1.0",
    "pgvector>=0.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
